[ray.llm] Refactor model download utilities (#51604)
