+++ python/ray/llm/_internal/batch/stages/tokenize_stage.py
@@ -6,10 +6,11 @@ from ray.llm._internal.batch.stages.base import (
     StatefulStage,
     StatefulStageUDF,
 )
-from ray.llm._internal.batch.utils import (
-    get_cached_tokenizer,
-    download_hf_model,
+from ray.llm._internal.common.utils.download_utils import (
+    download_model_files,
+    NodeModelDownloadable,
 )
+from ray.llm._internal.batch.utils import get_cached_tokenizer
 
 
 class TokenizeUDF(StatefulStageUDF):
@@ -30,7 +31,12 @@ class TokenizeUDF(StatefulStageUDF):
         from transformers import AutoTokenizer
 
         super().__init__(data_column, expected_input_keys)
-        model_path = download_hf_model(model, tokenizer_only=True)
+        model_path = download_model_files(
+            model_id=model,
+            mirror_config=None,
+            download_model=NodeModelDownloadable.TOKENIZER_ONLY,
+            download_extra_files=False,
+        )
         self.tokenizer = get_cached_tokenizer(
             AutoTokenizer.from_pretrained(
                 model_path,
@@ -88,7 +94,12 @@ class DetokenizeUDF(StatefulStageUDF):
         from transformers import AutoTokenizer
 
         super().__init__(data_column, expected_input_keys)
-        model_path = download_hf_model(model, tokenizer_only=True)
+        model_path = download_model_files(
+            model_id=model,
+            mirror_config=None,
+            download_model=NodeModelDownloadable.TOKENIZER_ONLY,
+            download_extra_files=False,
+        )
         self.tokenizer = get_cached_tokenizer(
             AutoTokenizer.from_pretrained(
                 model_path,
