+++ python/ray/llm/_internal/batch/utils.py
@@ -1,13 +1,6 @@
 """Utility functions for batch processing."""
 import logging
-import os
-from typing import TYPE_CHECKING, Any, Optional, Union
-
-from ray.llm._internal.common.utils.cloud_utils import (
-    CloudMirrorConfig,
-    is_remote_path,
-)
-from ray.llm._internal.common.utils.download_utils import CloudModelDownloader
+from typing import TYPE_CHECKING, Any, Union
 
 if TYPE_CHECKING:
     from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
@@ -67,51 +60,3 @@ def get_cached_tokenizer(tokenizer: AnyTokenizer) -> AnyTokenizer:
 
     tokenizer.__class__ = CachedTokenizer
     return tokenizer
-
-
-def download_hf_model(model_source: str, tokenizer_only: bool = True) -> str:
-    """Download the HF model from the model source.
-
-    Args:
-        model_source: The model source path.
-        tokenizer_only: Whether to download only the tokenizer.
-
-    Returns:
-        The local path to the downloaded model.
-    """
-
-    bucket_uri = None
-    if is_remote_path(model_source):
-        bucket_uri = model_source
-
-    mirror_config = CloudMirrorConfig(bucket_uri=bucket_uri)
-    downloader = CloudModelDownloader(model_source, mirror_config)
-    return downloader.get_model(tokenizer_only=tokenizer_only)
-
-
-def download_lora_adapter(
-    lora_name: str,
-    remote_path: Optional[str] = None,
-) -> str:
-    """If remote_path is specified, pull the lora to the local
-    directory and return the local path.
-
-    Args:
-        lora_name: The lora name.
-        remote_path: The remote path to the lora. If specified, the remote_path will be
-            used as the base path to load the lora.
-
-    Returns:
-        The local path to the lora if remote_path is specified, otherwise the lora name.
-    """
-    assert not is_remote_path(
-        lora_name
-    ), "lora_name cannot be a remote path (s3:// or gs://)"
-
-    if remote_path is None:
-        return lora_name
-
-    lora_path = os.path.join(remote_path, lora_name)
-    mirror_config = CloudMirrorConfig(bucket_uri=lora_path)
-    downloader = CloudModelDownloader(lora_name, mirror_config)
-    return downloader.get_model(tokenizer_only=False)
