+++ python/ray/llm/_internal/common/utils/cloud_utils.py
@@ -19,7 +19,7 @@ import asyncio
 # Use pyarrow for cloud storage access
 import pyarrow.fs as pa_fs
 
-from ray.llm._internal.serve.observability.logging import get_logger
+from ray.llm._internal.common.observability.logging import get_logger
 from ray.llm._internal.common.base_pydantic import BaseModelExtended
